fix: 결제 프로세스 수정 및 영수증 출력 모달 표시
