+++ src/app/(device-required)/pos/_components/modals/PayAlert.tsx
@@ -1,26 +1,22 @@
 "use client";
 
-import { useForm, Controller } from "react-hook-form";
+import { useForm } from "react-hook-form";
 import dynamic from "next/dynamic";
-import Button from "@/components/common/Button/Button";
-import Dropdown from "@/components/common/Dropdown";
-import Input from "@/components/common/Input";
-import Label from "@/components/common/Label";
 import { useRouter } from "next/navigation";
-import phoneNumberPattern from "@/lib/formatting/formatPhoneNumber";
 import { useDeviceContext } from "@/providers/deviceStoreProvider";
 import { zodResolver } from "@hookform/resolvers/zod";
-import { Form, FormErrorMessage } from "@/components/common/Form";
-import cn from "@/lib/utils";
-import formatLicenseNumber from "@/lib/formatting/formatLicenseNumber";
 import getQueryClient from "@/app/get-query-client";
 import { useState } from "react";
+import useOverlay from "@/hooks/useOverlay";
+import { Form } from "@/components/common/Form";
 import usePayment from "../../_queries/usePayment";
 import { print } from "../../_utils/print-receipt";
 import { useSelectItemStore } from "../../_hooks/useSelectItemStore";
 import { posQueries } from "../../_queries/usePos";
 import { paySchema, TypePayForm } from "../../_schema/pos.schema";
 import { posKeys } from "../../_queries/keys";
+import ReceiptModal from "./ReceiptModal";
+import PayAlertForm from "./PayAlertForm";
 
 const Alert = dynamic(() => import("@/components/common/Alert/Alert"), {
   ssr: false,
@@ -32,7 +28,7 @@ declare global {
   }
 }
 
-enum ReceiptType {
+export enum ReceiptType {
   NONE = "신청안함",
   DEDUCTION = "개인소득공제용",
   PROOF = "사업자증빙용",
@@ -54,10 +50,6 @@ export default function PayAlert({ close, type, ...props }: IProps) {
   const hasOrderId = selectedOrder?.orderId;
 
   const [isSubmitting, setIsSubmitting] = useState(false);
-
-  const menus = hasOrderId
-    ? selectedOrder?.orderMenus.map((el) => el.name)
-    : props?.orders.map((el) => el.orderMenus.map((v) => v.name)).flat();
 
   const selectedOrdersTotal = (selectedOrder?.orderMenus ?? []).reduce(
     (acc, menu) => {
@@ -94,10 +86,62 @@ export default function PayAlert({ close, type, ...props }: IProps) {
   const { data: activityData } = posQueries.useActivity(props.tableNo);
   const { data: stores } = posQueries.useStoreInfo(storeId!);
 
-  const monthlyPlan = [
-    "일시불",
-    ...new Array(11).fill(0).map((_, i) => (i + 2).toString().padStart(2, "0")),
-  ];
+  const receiptOverlay = useOverlay();
+
+  const handlePrintCard = (res: PaymentResponse) => {
+    print({
+      type: "card-receipt",
+      activity: activityData!,
+      stores: stores!,
+      payment: { ...res, INSTALLMENT: form.watch("monthlyPlan") },
+      successHandler: () => {
+        close();
+        if (props.orders?.length === 0) {
+          navigate.push("/pos/tables");
+        }
+      },
+    });
+  };
+
+  const handlePrintCash = () => {
+    print({
+      type: "cash-receipt",
+      activity: activityData!,
+      stores: stores!,
+      successHandler: () => setIsSubmitting(false),
+      cashReceiptPhoneNo: form.watch("phoneNumber"),
+    });
+  };
+
+  const handleModal = (res?: PaymentResponse) => {
+    receiptOverlay.open(() => (
+      <ReceiptModal
+        close={receiptOverlay.close}
+        onConfirm={() => {
+          if (type === "credit-card") {
+            handlePrintCard(res!);
+          } else {
+            handlePrintCash();
+          }
+          queryClient.invalidateQueries({
+            queryKey: posKeys.activity(props.tableNo),
+          });
+        }}
+        onCancel={() => {
+          receiptOverlay.close();
+          queryClient.invalidateQueries({
+            queryKey: posKeys.activity(props.tableNo),
+          });
+        }}
+      />
+    ));
+  };
+
+  const cashReceiptType = () => {
+    if (form.watch("receiptType") === ReceiptType.NONE) return "NONE";
+    if (form.watch("receiptType") === ReceiptType.PROOF) return "PROOF";
+    return "DEDUCTION";
+  };
 
   const handlePayment = () => {
     setIsSubmitting(true);
@@ -115,29 +159,12 @@ export default function PayAlert({ close, type, ...props }: IProps) {
         form,
         amount,
         successHandler: (res) => {
-          print({
-            type: "card-receipt",
-            activity: activityData!,
-            stores: stores!,
-            payment: { ...res, INSTALLMENT: form.watch("monthlyPlan") },
-            successHandler:
-              props.orders?.length > 0
-                ? () => close()
-                : () => {
-                    close();
-                    navigate.push("/pos/tables");
-                  },
-          });
+          close();
+          handleModal(res);
         },
+        errorHandler: () => setIsSubmitting(false),
       });
     } else {
-      let cashReceiptType = "";
-      if (form.watch("receiptType") === ReceiptType.NONE)
-        cashReceiptType = "NONE";
-      else if (form.watch("receiptType") === ReceiptType.PROOF)
-        cashReceiptType = "PROOF";
-      else cashReceiptType = "DEDUCTION";
-
       payCash({
         tableNo: props.tableNo,
         body: {
@@ -146,23 +173,11 @@ export default function PayAlert({ close, type, ...props }: IProps) {
             (form.watch("receiptType") === ReceiptType.PROOF
               ? form.watch("licenseNumber")
               : form.watch("phoneNumber")) ?? "",
-          cashReceiptType: cashReceiptType as OrderReceiptType,
+          cashReceiptType: cashReceiptType() as OrderReceiptType,
         },
         successHandler: () => {
-          print({
-            type: "cash-receipt",
-            activity: activityData!,
-            stores: stores!,
-            successHandler: () => {
-              close();
-              queryClient.invalidateQueries({
-                queryKey: posKeys.activity(props.tableNo),
-              });
-              navigate.push("/pos/tables");
-              setIsSubmitting(false);
-            },
-            cashReceiptPhoneNo: form.watch("phoneNumber"),
-          });
+          close();
+          handleModal();
         },
       });
     }
@@ -180,159 +195,12 @@ export default function PayAlert({ close, type, ...props }: IProps) {
       isSubmitted={isSubmitting}
     >
       <Form {...form}>
-        <div className="-mt-4 flex w-full flex-col gap-10">
-          <div className="flex items-center justify-between">
-            <h3 className="text-[28px] font-semibold">
-              {props.tableNo}번 테이블
-            </h3>
-            <Button
-              variant="outline"
-              color="primary"
-              className="button-lg !rounded-lg text-[15px] !font-medium"
-            >
-              결제 취소
-            </Button>
-          </div>
-          <div className="flex flex-col gap-8">
-            <div className="flex flex-col items-start">
-              <Label className="text-[15px] font-medium">결제 정보</Label>
-              <strong className="mt-2 text-2xl font-semibold">
-                {menus && (
-                  <strong className="mt-2 text-2xl font-semibold">
-                    {menus.length === 1
-                      ? menus[0]
-                      : `${menus[0]} 외 ${menus.length - 1}개`}
-                  </strong>
-                )}
-              </strong>
-            </div>
-            <div className="flex flex-col items-start">
-              <Label className="text-[15px] font-medium">결제할 금액</Label>
-              <strong className="mt-2 text-2xl font-semibold">
-                {(hasOrderId
-                  ? selectedOrdersTotal
-                  : props.remainingPaymentPrice
-                ).toLocaleString()}
-                원
-              </strong>
-            </div>
-            {type === "cash" ? (
-              <>
-                <div className="flex flex-col items-start">
-                  <Label className="text-[15px] font-medium">
-                    현금영수증 발행
-                  </Label>
-                  <div className="mt-2 flex w-full items-center gap-3">
-                    {Object.values(ReceiptType).map((key) => (
-                      <Button
-                        key={key}
-                        color={
-                          form.watch("receiptType") === key ? "primary" : "grey"
-                        }
-                        variant="outline"
-                        className={cn(
-                          "button-lg w-full !font-medium",
-                          form.watch("receiptType") === key
-                            ? ""
-                            : "border-gray-500"
-                        )}
-                        onClick={() => {
-                          form.setValue(
-                            "receiptType",
-                            key as TypePayForm["receiptType"]
-                          );
-                          form.setValue("phoneNumber", "");
-                          form.setValue("licenseNumber", "");
-                          form.clearErrors("phoneNumber");
-                          form.clearErrors("licenseNumber");
-                        }}
-                      >
-                        {key}
-                      </Button>
-                    ))}
-                  </div>
-                </div>
-                {form.watch("receiptType") === ReceiptType.DEDUCTION && (
-                  <div className="flex flex-col items-start gap-2">
-                    <Label className="text-[15px] font-medium">
-                      휴대폰 번호
-                    </Label>
-                    <Controller
-                      control={form.control}
-                      name="phoneNumber"
-                      render={({ field }) => (
-                        <div className="flex w-full flex-col gap-1">
-                          <Input
-                            {...field}
-                            placeholder="휴대폰 번호를 입력해주세요."
-                            className="w-full font-medium md:h-12"
-                            autoFocus
-                            onChange={(e) => {
-                              const onlyNums = e.target.value.replace(
-                                /[^0-9]/g,
-                                ""
-                              );
-                              const formatted = phoneNumberPattern(onlyNums);
-                              field.onChange(formatted);
-                            }}
-                            hasError={!!form.formState.errors.phoneNumber}
-                          />
-                          <FormErrorMessage>
-                            {form.formState.errors.phoneNumber?.message}
-                          </FormErrorMessage>
-                        </div>
-                      )}
-                    />
-                  </div>
-                )}
-                {form.watch("receiptType") === ReceiptType.PROOF && (
-                  <div className="flex flex-col items-start gap-2">
-                    <Label className="text-[15px] font-medium">
-                      사업자 번호
-                    </Label>
-                    <Controller
-                      control={form.control}
-                      name="licenseNumber"
-                      render={({ field }) => (
-                        <div className="flex w-full flex-col gap-1">
-                          <Input
-                            {...field}
-                            placeholder="사업자 번호를 입력해주세요."
-                            className="w-full font-medium md:h-12"
-                            autoFocus
-                            onChange={(e) => {
-                              const onlyNums = e.target.value.replace(
-                                /[^0-9]/g,
-                                ""
-                              );
-                              const formatted = formatLicenseNumber(onlyNums);
-                              field.onChange(formatted);
-                            }}
-                            hasError={!!form.formState.errors.licenseNumber}
-                          />
-                          <FormErrorMessage>
-                            {form.formState.errors.licenseNumber?.message}
-                          </FormErrorMessage>
-                        </div>
-                      )}
-                    />
-                  </div>
-                )}
-              </>
-            ) : (
-              <div className="flex flex-col items-start gap-2">
-                <Label className="text-[15px] font-medium">할부 개월</Label>
-                <Dropdown
-                  data={monthlyPlan}
-                  defaultText="할부 개월을 선택해주세요."
-                  active={form.watch("monthlyPlan").toString()}
-                  setActive={(value) => form.setValue("monthlyPlan", value)}
-                  triggerClassName="text-sm font-medium rounded-xl"
-                />
-              </div>
-            )}
-          </div>
-        </div>
+        <PayAlertForm
+          type={type}
+          hasOrderId={hasOrderId}
+          selectedOrdersTotal={selectedOrdersTotal}
+          {...props}
+        />
       </Form>
     </Alert>
   );
