+++ src/app/(device-required)/pos/_components/modals/ReceiptModal.tsx
@@ -1,6 +1,4 @@
 import dynamic from "next/dynamic";
-import usePayment from "../../_queries/usePayment";
-import { posQueries } from "../../_queries/usePos";
 
 const Alert = dynamic(() => import("@/components/common/Alert/Alert"), {
   ssr: false,
@@ -8,26 +6,25 @@ const Alert = dynamic(() => import("@/components/common/Alert/Alert"), {
 
 interface IProps {
   close: () => void;
-  tableNo: number;
-  storeId: string;
+  onConfirm: () => void;
+  onCancel?: () => void;
 }
 
-export default function ReceiptModal({ close, tableNo, storeId }: IProps) {
-  const { printReceipt } = usePayment();
-  const { data: activityData } = posQueries.useActivity(tableNo);
-  const { data: stores } = posQueries.useStoreInfo(storeId);
+export default function ReceiptModal({ close, onConfirm, onCancel }: IProps) {
+  const handleCancel = () => {
+    onCancel?.();
+    close();
+  };
 
-  const handleAction = () => {
-    printReceipt({
-      activity: activityData!,
-      stores: stores!,
-      successHandler: () => close(),
-    });
+  const handleConfirm = () => {
+    onConfirm();
+    close();
   };
 
   return (
     <Alert
-      onAction={handleAction}
+      onAction={handleConfirm}
+      onCancel={handleCancel}
       onClose={close}
       buttonText="출력하기"
       buttonColor="black"
